requirements.txt
quart>=0.19.0
hypercorn>=0.16.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
quart-cors>=0.7.0
Pillow>=10.0.0
python-multipart>=0.0.9
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
